arrange module
