+++ react-minesweeper/src/app/common.tsx
@@ -1,0 +1,13 @@
+
+export type MineBoardElement = {
+  isBomb: boolean;
+  isOpen: boolean;
+  bombCount: number;
+}
+
+export enum gameStatus {
+  ready,
+  inGame,
+  win,
+  loss,
+}
